+++ src/Modules/Admin/Widgets/Navs/EntryHeader.php
@@ -5,25 +5,19 @@ declare(strict_types=1);
 namespace Hirtz\Cms\Modules\Admin\Widgets\Navs;
 
 use Hirtz\Cms\Models\Entry;
-use Hirtz\Cms\Models\Section;
 use Hirtz\Cms\Modules\Admin\Data\EntryActiveDataProvider;
 use Hirtz\Cms\Modules\Admin\Widgets\Buttons\EntryCreateButton;
-use Hirtz\Cms\Modules\Admin\Widgets\Navs\Traits\EntryHeaderTrait;
-use Hirtz\Skeleton\Widgets\Navs\Header;
-use Hirtz\Skeleton\Widgets\Traits\ModelTrait;
+use Hirtz\Skeleton\Widgets\Navs\ModelHeader;
 use Hirtz\Skeleton\Widgets\Traits\ProviderTrait;
 use Override;
 use Stringable;
 use Yii;
 
-class EntryHeader extends Header
+/**
+ * @extends ModelHeader<Entry|null>
+ */
+class EntryHeader extends ModelHeader
 {
-    /**
-     * @use ModelTrait<Entry>
-     */
-    use ModelTrait;
-    use EntryHeaderTrait;
-
     /**
      * @use ProviderTrait<EntryActiveDataProvider|null>
      */
@@ -37,9 +31,6 @@ class EntryHeader extends Header
         if ($this->model) {
             $this->title ??= $this->model->getOldAttribute($this->model->getI18nAttributeName('name'));
             $this->subheading ??= FrontendLink::make()->model($this->model)->addClass('hidden-sticky');
-            $this->url ??= $this->model->getAdminRoute() ?: null;
-
-            $this->addEntryBreadcrumbs($this->model);
         }
 
         if ($this->provider) {