+++ src/Modules/Admin/Widgets/Navs/SectionSubmenu.php
@@ -41,22 +41,12 @@ class SectionSubmenu extends Submenu
     protected function configure(): void
     {
         $this->addItem(
-            $this->getSectionsItem(),
             $this->getSectionUpdateItem(),
             $this->getEntriesItem(),
             $this->getAssetsItem(),
         );
 
         parent::configure();
-    }
-
-    protected function getSectionsItem(): ?NavItem
-    {
-        return NavItem::make()
-            ->icon('angle-double-left')
-            ->badge($this->model->entry->section_count)
-            ->label(Yii::t('cms', 'COMMON_SECTIONS'))
-            ->url(['/admin/cms/section/index', 'entry' => $this->model->entry_id]);
     }
 
     protected function getSectionUpdateItem(): ?NavItem